move fish to /lib/
